EavFilesystemStorage without actor, but Arc<RwLock<>> instead
