--- cas_implementations/src/eav/file/mod.rs
@@ -1,103 +1,0 @@
-pub mod actor;
-use actor::{AskSelf, Protocol};
-use eav::file::actor::EavFileStorageActor;
-use holochain_core_types::{
-    eav::{Attribute, Entity, EntityAttributeValue, EntityAttributeValueStorage, Value},
-    error::{HcResult, HolochainError},
-};
-use riker::actors::*;
-use std::collections::HashSet;
-
-#[derive(Clone, PartialEq, Debug)]
-pub struct EavFileStorage {
-    actor: ActorRef<Protocol>,
-}
-
-impl EavFileStorage {
-    pub fn new(dir_path: String) -> HcResult<EavFileStorage> {
-        Ok(EavFileStorage {
-            actor: EavFileStorageActor::new_ref(&dir_path)?,
-        })
-    }
-}
-
-impl EntityAttributeValueStorage for EavFileStorage {
-    fn add_eav(&mut self, eav: &EntityAttributeValue) -> Result<(), HolochainError> {
-        let response = self.actor.block_on_ask(Protocol::EavAdd(eav.clone()))?;
-
-        match response {
-            Protocol::EavAddResult(add_result) => add_result,
-            _ => Err(HolochainError::ErrorGeneric(format!(
-                "Expected Protocol::EavAddResult received {:?}",
-                response
-            ))),
-        }
-    }
-
-    fn fetch_eav(
-        &self,
-        entity: Option<Entity>,
-        attribute: Option<Attribute>,
-        value: Option<Value>,
-    ) -> Result<HashSet<EntityAttributeValue>, HolochainError> {
-        let response = self
-            .actor
-            .block_on_ask(Protocol::EavFetch(entity, attribute, value))?;
-
-        match response {
-            Protocol::EavFetchResult(fetch_result) => fetch_result,
-            _ => Err(HolochainError::ErrorGeneric(format!(
-                "Expected Protocol::EavFetchResult received {:?}",
-                response
-            ))),
-        }
-    }
-}
-
-#[cfg(test)]
-pub mod tests {
-    extern crate tempfile;
-    use self::tempfile::tempdir;
-    use eav::file::EavFileStorage;
-    use holochain_core_types::{
-        cas::{
-            content::{AddressableContent, ExampleAddressableContent},
-            storage::EavTestSuite,
-        },
-        json::RawString,
-    };
-
-    #[test]
-    fn file_eav_round_trip() {
-        let temp = tempdir().expect("test was supposed to create temp dir");
-        let temp_path = String::from(temp.path().to_str().expect("temp dir could not be string"));
-        let entity_content =
-            ExampleAddressableContent::try_from_content(&RawString::from("foo").into()).unwrap();
-        let attribute = "favourite-color".to_string();
-        let value_content =
-            ExampleAddressableContent::try_from_content(&RawString::from("blue").into()).unwrap();
-        EavTestSuite::test_round_trip(
-            EavFileStorage::new(temp_path).unwrap(),
-            entity_content,
-            attribute,
-            value_content,
-        )
-    }
-
-    #[test]
-    fn file_eav_one_to_many() {
-        let temp = tempdir().expect("test was supposed to create temp dir");
-        let temp_path = String::from(temp.path().to_str().expect("temp dir could not be string"));
-        let eav_storage = EavFileStorage::new(temp_path).unwrap();
-        EavTestSuite::test_one_to_many::<ExampleAddressableContent, EavFileStorage>(eav_storage)
-    }
-
-    #[test]
-    fn file_eav_many_to_one() {
-        let temp = tempdir().expect("test was supposed to create temp dir");
-        let temp_path = String::from(temp.path().to_str().expect("temp dir could not be string"));
-        let eav_storage = EavFileStorage::new(temp_path).unwrap();
-        EavTestSuite::test_many_to_one::<ExampleAddressableContent, EavFileStorage>(eav_storage)
-    }
-
-}
